# backend/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Workload Tracker API",
        "version": "0.1.0",
        "description": "OpenAPI schema for Workload Tracker (People, Projects, Assignments, Deliverables, Departments)."
    },
    "paths": {
...
        "/api/jobs/{job_id}/": {
            "get": {
                "operationId": "jobs_retrieve",
                "description": "Return status and metadata for a Celery job.\n\nResponse fields:\n- id: task id\n- state: PENDING|STARTED|PROGRESS|SUCCESS|FAILURE\n- progress: 0-100 if available\n- message: optional status message\n- downloadReady: bool\n- downloadUrl: present when a file is available to download\n- result: task result when not file-based (e.g., import summary)\n- error: error message if failed",
                "parameters": [
                    {
                        "in": "path",
                        "name": "job_id",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Background job id",
                        "required": true
                    }
                ],
                "tags": [
                    "jobs"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "state": {
                                            "type": "string",
                                            "enum": [
                                                "PENDING",
                                                "STARTED",
                                                "PROGRESS",
                                                "SUCCESS",
                                                "FAILURE"
                                            ]
                                        },
                                        "progress": {
                                            "type": "integer"
                                        },
                                        "message": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "downloadReady": {
                                            "type": "boolean"
                                        },
                                        "downloadUrl": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "result": {},
                                        "error": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "required": [
                                        "id",
                                        "state",
                                        "progress",
                                        "downloadReady"
                                    ]
                                }
                            }
                        },
                        "description": ""
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Async jobs not available"
                                }
                            }
                        },
                        "description": ""
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Job not completed"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/jobs/{job_id}/download/": {
            "get": {
                "operationId": "jobs_download_retrieve",
                "description": "Stream the file produced by a completed job (if any).",
                "parameters": [
                    {
                        "in": "path",
                        "name": "job_id",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Background job id",
                        "required": true
                    }
                ],
                "tags": [
                    "jobs"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        },
                        "description": "File content"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "No file associated with this job"
                                }
                            }
                        },
                        "description": ""
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Job not completed"
                                }
                            }
                        },
                        "description": ""
                    },
                    "503": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Async jobs not available"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sessionid"
            },
            "jwtAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}